saradc: the sampling data maybe wrong when adjust the clock [1/1]
PD#SWPL-5566

Problem:
If we change the clock to 24M/(20+1), then use the command below
to sample, and the sampling data maybe wrong.

- saradc open 1; saradc getval; saradc close

Solution:
through lots of tests, as long as we avoid sudden changing of REG3
bit[27] during initialization (1->0->1), the sampling data is corrent.

Verify:
test pass on txhd_skt

Change-Id: I2722ddb4b48c1eec46b4d7ab88b48edd725d41a5
Signed-off-by: Xingyu Chen <xingyu.chen@amlogic.com>